LayoutTests/http/tests/iframe-monitor/cached-resource-expected.txt

Cached huge resource is not handled as network usage.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is 20480
PASS document.querySelector('iframe[name=frame1]').srcdoc is ""
PASS successfullyParsed is true

TEST COMPLETE
